Adicionado arquitetura armeabi-v7a

+++ ariesframework/src/main/java/org/hyperledger/ariesframework/agent/AgentConfig.kt
@@ -18,7 +18,7 @@ enum class MediatorPickupStrategy {
 
 @Serializable
 data class BesuLedgerConfig(
-    val chainId: Int,
+    val chainId: ULong,
     val nodeAddress: String,
     val network: String? = null,
 )
